Combine sample types into shared interface
