+++ src/faintLinkDef.h
@@ -4,8 +4,7 @@
 #include "faint/PreSelection.h"
 #include "faint/Run.h"
 #include "faint/RunReader.h"
-#include "faint/Sample.h"
-#include "faint/SampleSet.h"
+#include "faint/Samples.h"
 #include "faint/SelectionQuery.h"
 #include "faint/plot/StackedHistogram.h"
 #include "faint/TruthClassifier.h"
